Move regression tests into separate folder
