+++ tests/regression/tga/bad-color-map.tga
