+++ tests/regression/tga/flip-zero-height.tga
