+++ tests/regression/tga/sixteen-bit-gray-colormap.tga
