+++ tests/regression/tga/sixteen-bit-gray-colormap2.tga
